requirements.txt
django==1.9.1
djangorestframework==3.6.2
